pyproject.toml
[tool.poetry]
name = "medgpt-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.0"
loguru = "^0.7.2"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
aiohttp = "^3.10.9"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-multipart = "^0.0.12"
websockets = "^13.1"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
